# contract/openapi.yaml
openapi: 3.0.0
paths:
...
  /v1/documents:
    post:
      description: Requires documents:create. Stores a stubbed original_file_url;
        status draft, owner = caller.
      operationId: DocumentsController_upload_v1
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDocumentDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentResponse"
        "400": &a155
          description: Bad Request — validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "401": &a156
          description: Unauthorized — missing/invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "403": &a157
          description: Forbidden — RBAC denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "404": &a158
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "409": &a159
          description: Conflict — invalid state transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "422": &a160
          description: Unprocessable Entity — domain rule violated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
      security: &a161
        - bearer: []
      summary: Upload a document
      tags: &a162
        - Documents & E-Signature
    get:
      description: Requires documents:view. Scoped to owned or shared-with
        (Admin/Super Admin see all).
      operationId: DocumentsController_list_v1
      parameters:
        - name: status
          required: false
          in: query
          schema:
            type: string
            enum:
              - draft
              - shared
              - partially_signed
              - completed
              - declined
        - name: doc_type
          required: false
          in: query
          schema:
            type: string
            enum:
              - compensation_agreement
              - rate_notice
              - equipment
              - other
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentResponse"
        "400": *a155
        "401": *a156
        "403": *a157
        "404": *a158
        "409": *a159
        "422": *a160
      security: *a161
      summary: List documents
      tags: *a162
  /v1/documents/{id}:
    get:
      description: Requires documents:view. 404 if not visible to the caller.
      operationId: DocumentsController_findOne_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentResponse"
        "400": *a155
        "401": *a156
        "403": *a157
        "404": *a158
        "409": *a159
        "422": *a160
      security: *a161
      summary: Get a document (requests + per-signer status)
      tags: *a162
  /v1/documents/{id}/signature-requests:
    post:
      description: Requires documents:create AND ownership of the document (else 403).
        Recipients become the shared-with set.
      operationId: DocumentsController_requestSignature_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSignatureRequestDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignatureRequestResponse"
        "400": *a155
        "401": *a156
        "403": *a157
        "404": *a158
        "409": *a159
        "422": *a160
      security: *a161
      summary: Share + request a signature from one or many recipients
      tags: *a162
  /v1/signature-requests/{id}/sign:
    post:
      description: Authenticated; the caller must be a recipient of this request (else
        403). decision = sign | decline. Records the signature event (method,
        IP, timestamp) and stores a stubbed signed copy.
      operationId: SignatureRequestsController_sign_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignActionResultResponse"
        "400": &a163
          description: Bad Request — validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "401": &a164
          description: Unauthorized — missing/invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "403": &a165
          description: Forbidden — RBAC denied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "404": &a166
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "409": &a167
          description: Conflict — invalid state transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
        "422": &a168
          description: Unprocessable Entity — domain rule violated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelopeDto"
      security: &a169
        - bearer: []
      summary: Sign or decline a signature request
      tags: &a170
        - Documents & E-Signature
  /v1/signature-requests/{id}/cancel:
    post:
      description: Authenticated; only the requester, the document owner, or an admin
        may cancel (else 403).
      operationId: SignatureRequestsController_cancel_v1
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CancelSignatureResultResponse"
        "400": *a163
        "401": *a164
        "403": *a165
        "404": *a166
        "409": *a167
        "422": *a168
      security: *a169
      summary: Cancel a signature request
      tags: *a170
...
info:
  title: Redwave ERP / HRM API
  description: Versioned REST contract for the Redwave platform. Each protected
    endpoint notes the required (module, action) permission enforced server-side
    by the RBAC guard.
  version: "1.0"
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    ErrorBodyDto:
      type: object
      properties:
        code:
          type: string
          example: UNPROCESSABLE_ENTITY
          description: Stable machine code for the error.
        message:
          type: string
          example: tier brackets must be contiguous (each min_count = previous max_count +
            1)
        details:
          type: object
          description: Optional structured detail (e.g. validation messages, an
            unpriced-products list).
          additionalProperties: true
      required:
        - code
        - message
    ErrorEnvelopeDto:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/ErrorBodyDto"
      required:
        - error
...
    CreateDocumentDto:
      type: object
      properties:
        title:
          type: string
          example: Compensation Agreement 2026
        doc_type:
          type: string
          enum:
            - compensation_agreement
            - rate_notice
            - equipment
            - other
          example: compensation_agreement
      required:
        - title
        - doc_type
    DocumentSignatureResponse:
      type: object
      properties:
        id:
          type: string
        recipient_user_id:
          type: string
        status:
          type: string
          enum:
            - pending
            - signed
            - declined
        signed_file_url:
          type: string
          nullable: true
          description: Per-signer signed copy (stub).
        signed_at:
          type: string
          format: date-time
          nullable: true
        method:
          type: string
          nullable: true
          description: e.g. click_to_sign.
      required:
        - id
        - recipient_user_id
        - status
        - signed_file_url
        - signed_at
        - method
    SignatureRequestResponse:
      type: object
      properties:
        id:
          type: string
        document_id:
          type: string
        requested_by:
          type: string
        message:
          type: string
          nullable: true
        due_date:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum:
            - pending
            - completed
            - declined
            - cancelled
        created_at:
          type: string
          format: date-time
        document_signatures:
          type: array
          items:
            $ref: "#/components/schemas/DocumentSignatureResponse"
      required:
        - id
        - document_id
        - requested_by
        - message
        - due_date
        - status
        - created_at
        - document_signatures
    DocumentResponse:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        doc_type:
          type: string
          enum:
            - compensation_agreement
            - rate_notice
            - equipment
            - other
        owner_user_id:
          type: string
        original_file_url:
          type: string
          description: Stub storage reference; never mutated (DOC-004).
        status:
          type: string
          enum:
            - draft
            - shared
            - partially_signed
            - completed
            - declined
        created_at:
          type: string
          format: date-time
        signature_requests:
          description: Present on the detail endpoint; absent on the list.
          type: array
          items:
            $ref: "#/components/schemas/SignatureRequestResponse"
      required:
        - id
        - title
        - doc_type
        - owner_user_id
        - original_file_url
        - status
        - created_at
    CreateSignatureRequestDto:
      type: object
      properties:
        recipient_user_ids:
          description: User ids to request a signature from (≥1).
          type: array
          items:
            type: string
        message:
          type: string
          example: Please review and sign by Friday.
        due_date:
          type: string
          example: 2026-06-30
          description: Optional due date (record only).
      required:
        - recipient_user_ids
    SignDto:
      type: object
      properties:
        decision:
          type: string
          enum:
            - sign
            - decline
        method:
          type: string
          example: typed
          description: Signature method (typed / drawn / …).
      required:
        - decision
    SignActionResultResponse:
      type: object
      properties:
        signatureId:
          type: string
        documentId:
          type: string
        ownerId:
          type: string
        requestStatus:
          type: string
          enum:
            - pending
            - completed
            - declined
            - cancelled
        documentStatus:
          type: string
          enum:
            - draft
            - shared
            - partially_signed
            - completed
            - declined
      required:
        - signatureId
        - documentId
        - ownerId
        - requestStatus
        - documentStatus
    CancelSignatureResultResponse:
      type: object
      properties:
        request_id:
          type: string
        status:
          type: string
          enum:
            - pending
            - completed
            - declined
            - cancelled
          example: cancelled
        document_status:
          type: string
          enum:
            - draft
            - shared
            - partially_signed
            - completed
            - declined
      required:
        - request_id
        - status
        - document_status
...